Demo_Users/XL/requirements.txt
gradio
requests
paho-mqtt
matplotlib
seaborn
pandas
numpy
datetime